+++ src/app/home/setting/expenseCategory/_components/ExpenseCategoryContainer.tsx
@@ -1,70 +1,14 @@
 import NoExpenseCategory from "./NoExpenseCategory";
 import ExpenseCategory from "./ExpenseCategory";
 import { ExpenseCategoryContainerProps } from "./ExpenseCategoryContainer.type";
-import useModalStore from "@/stores/modalStore";
-import UpdateExpenseCategoryModal from "./UpdateExpenseCategoryModal";
-import DeleteExpenseCategoryModal from "./DeleteExpenseCategoryModal";
+import useExpenseCategoryContainer from "./ExpenseCategoryContainer.hook";
 
 export default function ExpenseCategoryContainer({
     expenseCategories,
     expenseCategoriesDispatch,
 }: ExpenseCategoryContainerProps) {
-    const { setIsModalOpen } = useModalStore();
-
-    const handleUpdateCategoryModalOpenButtonClick = (e: React.MouseEvent) => {
-        const currentExpenseCategoryName = (
-            e.target as HTMLElement
-        ).closest<HTMLButtonElement>("button")?.dataset.name;
-
-        const currentExpenseCategoryId = Number(
-            (e.target as HTMLElement).closest<HTMLButtonElement>("button")
-                ?.dataset.id
-        );
-
-        if (!currentExpenseCategoryName || !currentExpenseCategoryId) return;
-
-        const currentExpenseCategory = {
-            id: currentExpenseCategoryId,
-            name: currentExpenseCategoryName,
-        };
-
-        setIsModalOpen(
-            <UpdateExpenseCategoryModal
-                currentExpenseCategory={currentExpenseCategory}
-                expenseCategoriesDispatch={expenseCategoriesDispatch}
-            />
-        );
-    };
-
-    const handleDeleteCategoryModalOpenButtonClick = (e: React.MouseEvent) => {
-        const currentExpenseCategoryId = Number(
-            (e.target as HTMLElement).closest<HTMLButtonElement>("button")
-                ?.dataset.id
-        );
-
-        if (!currentExpenseCategoryId) return;
-
-        setIsModalOpen(
-            <DeleteExpenseCategoryModal
-                currentExpenseCategoryId={currentExpenseCategoryId}
-                expenseCategoriesDispatch={expenseCategoriesDispatch}
-            />
-        );
-    };
-
-    const handleUpdateOrDeleteCategoryButtonClick = (e: React.MouseEvent) => {
-        const buttonType = (e.target as HTMLElement).closest<HTMLButtonElement>(
-            "button"
-        )?.dataset.type;
-
-        if (!buttonType) return;
-
-        if (buttonType === "update") {
-            handleUpdateCategoryModalOpenButtonClick(e);
-        } else if (buttonType === "delete") {
-            handleDeleteCategoryModalOpenButtonClick(e);
-        }
-    };
+    const { handleUpdateOrDeleteCategoryButtonClick } =
+        useExpenseCategoryContainer({ expenseCategoriesDispatch });
 
     return (
         <ul
