+++ src/app/home/setting/expenseCategory/_components/ExpenseCategoryContainer.type.ts
@@ -5,3 +5,7 @@ export interface ExpenseCategoryContainerProps {
     expenseCategories: ExpenseCategoryInterface[];
     expenseCategoriesDispatch: React.Dispatch<ExpenseCategoryAction>;
 }
+
+export interface UseExpenseCategoryContainerProps {
+    expenseCategoriesDispatch: React.Dispatch<ExpenseCategoryAction>;
+}
